[filegen] move devcontainer_util.go into package filegen; make tmplFS internal
